% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Severity_methods_assessment-package.R
\docType{package}
\name{Severity_methods_assessment-package}
\alias{Severity_methods_assessment}
\alias{Severity_methods_assessment-package}
\title{Severity_methods_assessment: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CarmenTamayo/Severity_methods_assessment}
  \item Report bugs at \url{https://github.com/CarmenTamayo/Severity_methods_assessment/issues}
}

}
\author{
\strong{Maintainer}: Carmen Tamayo Cuartero \email{carmen.tamayo-cuartero@lshtm.ac.uk} (\href{https://orcid.org/0000-0003-4184-2864}{ORCID}) [copyright holder]

}
\keyword{internal}
